Extract tree-sitter colorizer backend
